Move button-related func. to separate viewModel
