+++ src/ggrc-client/js/components/assessment/assessments-bulk-complete-button/assessments-bulk-complete-button.js
@@ -3,15 +3,12 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import canMap from 'can-map';
 import canComponent from 'can-component';
 import canStache from 'can-stache';
 import template from './assessments-bulk-complete-button.stache';
-import {isMyAssessments} from '../../../plugins/utils/current-page-utils';
+import BulkUpdatableButton from '../view-models/bulk-updatable-button-vm';
 
-const viewModel = canMap.extend({
-  isButtonView: false,
-  parentInstance: null,
+const viewModel = BulkUpdatableButton.extend({
   async openBulkCompleteModal(el) {
     const {AssessmentsBulkComplete} = await import(
       /* webpackChunkName: "mapper" */
@@ -19,17 +16,6 @@ const viewModel = canMap.extend({
     );
 
     AssessmentsBulkComplete.launch($(el), this.getModalConfig());
-  },
-  getModalConfig() {
-    const parentInstance = this.attr('parentInstance');
-    return {
-      isMyAssessmentsView: isMyAssessments(),
-      mappedToItems: parentInstance ? [{
-        id: parentInstance.attr('id'),
-        type: parentInstance.attr('type'),
-        title: parentInstance.attr('title'),
-      }] : [],
-    };
   },
 });
