+++ src/ggrc-client/js/components/assessment/view-models/bulk-updatable-button-vm.js
@@ -1,0 +1,23 @@
+/*
+  Copyright (C) 2019 Google Inc.
+  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import canMap from 'can-map';
+import {isMyAssessments} from '../../../plugins/utils/current-page-utils';
+
+export default canMap.extend({
+  isButtonView: false,
+  parentInstance: null,
+  getModalConfig() {
+    const parentInstance = this.attr('parentInstance');
+    return {
+      isMyAssessmentsView: isMyAssessments(),
+      mappedToItems: parentInstance ? [{
+        id: parentInstance.attr('id'),
+        type: parentInstance.attr('type'),
+        title: parentInstance.attr('title'),
+      }] : [],
+    };
+  },
+});
